+++ mantis/__init__.py
@@ -62,7 +62,7 @@ from .discovery import (
 from .constants import *
 
 #: SDK 版本号
-__version__ = "1.3.10"
+__version__ = "1.3.11"
 
 #: 作者
 __author__ = "BlueWorm-EAI-Tech"